added (not 100% correct?) implementation of time_hzcrr

+++ pc/recog/src/recog.c
@@ -1,8 +1,8 @@
 #include "recog.h"
 
 uint8_t buffer[BUF_SIZE];
-uint8_t zcr[FRAMES_PER_SAMPLE];
-uint8_t ste[FRAMES_PER_SAMPLE];
+int zcr[FRAMES_PER_SAMPLE];
+int ste[FRAMES_PER_SAMPLE];
 
 void print_buffer(ssize_t bytesread)
 {
